+++ src/mlpack/tests/recurrent_network_test.cpp
@@ -1414,9 +1414,6 @@ BOOST_AUTO_TEST_CASE(BRNNTrainReturnObjective)
   BOOST_REQUIRE_EQUAL(std::isfinite(objVal), true);
 }
 
-/**
- * Test a BRNN model with PrintLoss callback and optimizer.
- */
 BOOST_AUTO_TEST_CASE(BRNNWithOptimizerCallbackTest)
 {
   const size_t rho = 10;
@@ -1453,9 +1450,6 @@ BOOST_AUTO_TEST_CASE(BRNNWithOptimizerCallbackTest)
   BOOST_REQUIRE_GT(stream.str().length(), 0);
 }
 
-/**
- * Test a BRNN model with PrintLoss callbacks.
- */
 BOOST_AUTO_TEST_CASE(BRNNCallbackTest)
 {
   const size_t rho = 10;
@@ -1491,5 +1485,64 @@ BOOST_AUTO_TEST_CASE(BRNNCallbackTest)
   BOOST_REQUIRE_GT(stream.str().length(), 0);
 }
 
+BOOST_AUTO_TEST_CASE(LargeRhoValueRnnTest)
+{
+  // Setting rho value greater than sequence length which is 17.
+  const size_t rho = 100;
+  const size_t hiddenSize = 128;
+  const size_t numLetters = 256;
+  using MatType = arma::cube;
+  std::vector<std::string>trainingData = { "THIS IS THE INPUT 0" ,
+                                           "THIS IS THE INPUT 1" ,
+                                           "THIS IS THE INPUT 3"};
+
+
+  RNN<> model(rho);
+  model.Add<IdentityLayer<>>();
+  model.Add<LSTM<>>(numLetters, hiddenSize, rho);
+  model.Add<Dropout<>>(0.1);
+  model.Add<Linear<>>(hiddenSize, numLetters);
+
+  const auto makeInput = [numLetters](const char *line) -> MatType
+  {
+    const auto strLen = strlen(line);
+    // Rows: number of dimensions.
+    // Cols: number of sequences/points.
+    // Slices: number of steps in sequences.
+    MatType result(numLetters, 1, strLen, arma::fill::zeros);
+    for (size_t i = 0; i < strLen; ++i)
+    {
+      result.at(static_cast<arma::uword>(line[i]), 0, i) = 1.0;
+    }
+    return result;
+  };
+
+  const auto makeTarget = [] (const char *line) -> MatType
+  {
+    const auto strLen = strlen(line);
+    // Responses for NegativeLogLikelihood should be
+    // non-one-hot-encoded class IDs (from 1 to num_classes).
+    MatType result(1, 1, strLen, arma::fill::zeros);
+    // The response is the *next* letter in the sequence.
+    for (size_t i = 0; i < strLen - 1; ++i)
+    {
+      result.at(0, 0, i) = static_cast<arma::uword>(line[i + 1]) + 1.0;
+    }
+    // The final response is empty, so we set it to class 0.
+    result.at(0, 0, strLen - 1) = 1.0;
+    return result;
+  };
+
+  std::vector<MatType> inputs(trainingData.size());
+  std::vector<MatType> targets(trainingData.size());
+  for (size_t i = 0; i < trainingData.size(); ++i)
+  {
+    inputs[i] = makeInput(trainingData[i].c_str());
+    targets[i] = makeTarget(trainingData[i].c_str());
+  }
+  ens::SGD<> opt(0.01, 1, 100);
+  double objVal = model.Train(inputs[0], targets[0], opt);
+  BOOST_TEST_CHECKPOINT("Training over");
+}
 
 BOOST_AUTO_TEST_SUITE_END();