+++ HISTORY.md
@@ -40,6 +40,8 @@
 
   * Add `__version__` to `__init__.py` (#2092).
 
+  * Correctly handle RNN sequences that are shorter than the value of rho (#2102).
+
 ### mlpack 3.2.1
 ###### 2019-10-01
   * Enforce CMake version check for ensmallen (#2032).